+++ Editor/InlineDrawer.cs
@@ -1,5 +1,6 @@
 using UnityEditor;
 using UnityEditor.UIElements;
+using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace Vertx.Attributes.Editor
@@ -7,13 +8,34 @@ namespace Vertx.Attributes.Editor
 	[CustomPropertyDrawer(typeof(InlineAttribute))]
 	public sealed class InlineDrawer : DecoratorDrawer
 	{
+		private sealed class InlineInspectorElement : InspectorElement
+		{
+			private readonly Object _target;
+
+			public InlineInspectorElement(SerializedProperty property) : base(property.objectReferenceValue)
+			{
+				_target = property.objectReferenceValue;
+				this.TrackPropertyValue(property, TrackChanges);
+			}
+
+			private void TrackChanges(SerializedProperty property)
+			{
+				if (property.objectReferenceValue == _target)
+					return;
+				VisualElement p = parent;
+				if (p == null)
+					return;
+				RemoveFromHierarchy();
+				p.Add(new InlineInspectorElement(property));
+			}
+		}
+		
 		public override VisualElement CreatePropertyGUI() => new DecoratePropertyElement((property, element) =>
 		{
 			if (property.propertyType == SerializedPropertyType.ObjectReference)
 			{
-				if (element.parent[element.parent.childCount - 1] is InspectorElement) return;
-				if (property.objectReferenceValue == null) return;
-				element.parent.Add(new InspectorElement(property.objectReferenceValue));
+				if (element.parent[element.parent.childCount - 1] is InlineInspectorElement) return;
+				element.parent.Add(new InlineInspectorElement(property));
 				return;
 			}