+++ Editor/Utilities/EnumFlagsValueAndNames.cs
@@ -10,6 +10,9 @@ using UnityEngine;
 
 namespace Vertx.Attributes.Editor
 {
+#if UNITY_6000_7_OR_NEWER
+	[Unity.Scripting.LifecycleManagement.NoAutoStaticsCleanup]
+#endif
 	internal sealed class EnumFlagsValueAndNames
 	{
 		private static readonly Dictionary<Type, EnumFlagsValueAndNames?> s_lookup = new Dictionary<Type, EnumFlagsValueAndNames?>();