+++ Runtime/PlayOnlyFieldAttribute.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace Vertx.Attributes
+{
+	public sealed class PlayOnlyFieldAttribute : PropertyAttribute { }
+}